Move the journal ffi to its own module.
Similar to how the rest of libsystemd-sys is organized, move the journal to
its own module.
