+++ src/journal.rs
@@ -2,7 +2,8 @@ use libc::{c_int, size_t};
 use log::{self, Log, LogRecord, LogLocation, SetLoggerError};
 use std::{fmt, ptr, result};
 use std::collections::BTreeMap;
-use ffi;
+use ffi::array_to_iovecs;
+use ffi::journal as ffi;
 use super::Result;
 
 /// Send preformatted fields to systemd.
@@ -10,7 +11,7 @@ use super::Result;
 /// This is a relatively low-level operation and probably not suitable unless
 /// you need precise control over which fields are sent to systemd.
 pub fn send(args: &[&str]) -> c_int {
-    let iovecs = ffi::array_to_iovecs(args);
+    let iovecs = array_to_iovecs(args);
     unsafe { ffi::sd_journal_sendv(iovecs.as_ptr(), iovecs.len() as c_int) }
 }
